Move injector to base class
