+++ app/src/main/java/com/mindorks/framework/mvvm/ui/base/BaseActivity.java
@@ -35,6 +35,7 @@ import com.mindorks.framework.mvvm.ui.login.LoginActivity;
 import com.mindorks.framework.mvvm.utils.CommonUtils;
 import com.mindorks.framework.mvvm.utils.NetworkUtils;
 
+import dagger.android.AndroidInjection;
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
 /**
@@ -156,7 +157,9 @@ public abstract class BaseActivity<T extends ViewDataBinding, V extends BaseView
     @LayoutRes
     int getLayoutId();
 
-    public abstract void performDependencyInjection();
+    public void performDependencyInjection() {
+        AndroidInjection.inject(this);
+    }
 
 }
