+++ app/src/main/java/com/mindorks/framework/mvvm/ui/login/LoginActivity.java
@@ -29,8 +29,6 @@ import com.mindorks.framework.mvvm.ui.main.MainActivity;
 
 import javax.inject.Inject;
 
-import dagger.android.AndroidInjection;
-
 /**
  * Created by amitshekhar on 08/07/17.
  */
@@ -97,11 +95,6 @@ public class LoginActivity extends BaseActivity<ActivityLoginBinding, LoginViewM
     @Override
     public int getLayoutId() {
         return R.layout.activity_login;
-    }
-
-    @Override
-    public void performDependencyInjection() {
-        AndroidInjection.inject(this);
     }
 
 }
