+++ app/src/main/java/com/mindorks/framework/mvvm/ui/main/MainActivity.java
@@ -59,7 +59,6 @@ import java.util.ArrayList;
 
 import javax.inject.Inject;
 
-import dagger.android.AndroidInjection;
 import dagger.android.AndroidInjector;
 import dagger.android.DispatchingAndroidInjector;
 import dagger.android.support.HasSupportFragmentInjector;
@@ -329,11 +328,6 @@ public class MainActivity extends BaseActivity<ActivityMainBinding, MainViewMode
     @Override
     public int getLayoutId() {
         return R.layout.activity_main;
-    }
-
-    @Override
-    public void performDependencyInjection() {
-        AndroidInjection.inject(this);
     }
 
     @Override
